Output_data_P9_run7_57_CutTester_test9

+++ src/cuts/VetoCuts.h
@@ -47,7 +47,7 @@ double ToF_n_ucut = 20.;
 double Edep_CND_lcut = 5.;
 
 /* Step 2 cuts (PARTIAL!) */
-double Cluster_size_cut = 1.;
-double CND1_LayerMult_cut = 1.;
-double CND2andCND3_LayerMult_ucut = 2.;
+int Cluster_size_cut = 1;
+int CND1_LayerMult_cut = 1;
+int CND2andCND3_LayerMult_ucut = 2;
 #endif //VETOCUTS_H
